+++ packages/building-rollup/index.js
@@ -3,8 +3,8 @@
  * @typedef {import('./src/types').SpaOptions} SpaOptions
  */
 
-const { createBasicConfig } = require('./src/createBasicConfig');
-const { createSpaConfig } = require('./src/createSpaConfig');
-const { createMpaConfig } = require('./src/createMpaConfig');
+const { createBasicConfig } = require('./src/createBasicConfig.js');
+const { createSpaConfig } = require('./src/createSpaConfig.js');
+const { createMpaConfig } = require('./src/createMpaConfig.js');
 
 module.exports = { createBasicConfig, createSpaConfig, createMpaConfig };