+++ packages/eslint-plugin-lit-a11y/lib/rules/tabindex-no-positive.js
@@ -19,8 +19,7 @@ const TabindexNoPositiveRule = {
       description: 'Enforce tabIndex value is not greater than zero.',
       category: 'Accessibility',
       recommended: false,
-      url:
-        'https://github.com/open-wc/open-wc/blob/master/packages/eslint-plugin-lit-a11y/docs/rules/tabindex-no-positive.md',
+      url: 'https://github.com/open-wc/open-wc/blob/master/packages/eslint-plugin-lit-a11y/docs/rules/tabindex-no-positive.md',
     },
     messages: {
       tabindexNoPositive: 'Invalid tabindex value {{value}}.',