+++ packages/eslint-plugin-lit-a11y/tests/lib/rules/role-has-required-aria-attrs.js
@@ -8,7 +8,7 @@
 //------------------------------------------------------------------------------
 
 const { RuleTester } = require('eslint');
-const rule = require('../../../lib/rules/role-has-required-aria-attrs');
+const rule = require('../../../lib/rules/role-has-required-aria-attrs.js');
 
 //------------------------------------------------------------------------------
 // Tests
@@ -26,19 +26,16 @@ ruleTester.run('role-has-required-aria-attrs', rule, {
   valid: [
     { code: "html`<span role='alert' aria-atomic='foo' aria-live='foo'></span>`" },
     {
-      code:
-        'html`<span role="checkbox" aria-checked="false" aria-labelledby="foo" tabindex="0"></span>`',
+      code: 'html`<span role="checkbox" aria-checked="false" aria-labelledby="foo" tabindex="0"></span>`',
     },
     { code: 'html`<span role="row"></span>`' },
     { code: 'html`<input type="checkbox" role="switch" aria-checked="true" />`' },
     { code: 'html`<div role="combobox" aria-controls="foo" aria-expanded="foo"></div>`' },
     {
-      code:
-        'html`<div role="combobox" aria-activedescendant="child" aria-autocomplete="none" aria-controls="listbox" aria-expanded="false"></div>`;',
+      code: 'html`<div role="combobox" aria-activedescendant="child" aria-autocomplete="none" aria-controls="listbox" aria-expanded="false"></div>`;',
     },
     {
-      code:
-        'html`<div role="combobox" aria-activedescendant=${"child"} aria-autocomplete=${"none"} aria-controls="listbox" aria-expanded=${"false"}></div>`;',
+      code: 'html`<div role="combobox" aria-activedescendant=${"child"} aria-autocomplete=${"none"} aria-controls="listbox" aria-expanded=${"false"}></div>`;',
     },
   ],