+++ packages/testing/index.d.ts
@@ -17,8 +17,9 @@ export { fixtureCleanup } from '@open-wc/testing-helpers';
 export { elementUpdated } from '@open-wc/testing-helpers';
 export { waitUntil } from '@open-wc/testing-helpers';
 
-import chai from 'chai';
+import Chai from 'chai';
 
-export declare const expect: typeof chai.expect;
-export declare const assert: typeof chai.assert;
-export declare const should: typeof chai.should;
+export declare const chai: typeof Chai;
+export declare const expect: typeof Chai.expect;
+export declare const assert: typeof Chai.assert;
+export declare const should: typeof Chai.should;