+++ packages/semantic-dom-diff/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 0.19.7
+
+### Patch Changes
+
+- 735b0092: Avoid saving unchanged snapshots
+
+## 0.19.6
+
+### Patch Changes
+
+- a7e9c20d: chore(semantic-dom-diff): updates dependencies
+
 ## 0.19.5
 
 ### Patch Changes